+++ example/lib/assets/example_assets.dart
@@ -1,4 +1,7 @@
 class ExampleAssets {
-  static const String mapPinImageEncoded =
-      'iVBORw0KGgoAAAANSUhEUgAAACAAAAAgCAYAAABzenr0AAAACXBIWXMAAADsAAAA7AF5KHG9AAAAGXRFWHRTb2Z0d2FyZQB3d3cuaW5rc2NhcGUub3Jnm+48GgAAA39JREFUWIXFl11oHFUUx39n7m66s7thJaaJ1aCxttYk9U3yYEGSWsxHBfHBvBYlFKuCpe9SUDCgL4oiJSBNfax98APTPCjb2voQQRDbZNOauhZaP5Iq2XY3m2Zn5vgQ+5WdmZ1NlJ7Hc//n/n/MnDn3DtzlkLUUtR1VO50qbbFE2gA81UvFUmr20pCU/1eAzuPlp1X1VaAPSK5aXkSZ8FQ+PPesnf1PAbZ/Xmx14+Yw6ECkXVW+qojz4uxg4/y6ATonrnWpZ8aBByOZ34qLFmZwanDD9JoBHv3yanPMxCYVNtdpfhMibtzun/oa54IEVli1MfFP1mEO8FDFNR+HCQKfwGPHyz2i6ttMO1sNe9pjbM+s8J8peBzJO2TnXH8TkV3TA/Y3fmuxQDLV1/3yB7bFGd58Z1l3k0V3UwOjFxzeO1+pLlLdB/gC+L6CtqNqA8+szu9sNVXmt8feR2L0tJhqf+hrz2oiMkAmVdpG9XfOnvZg8xqatF0qbokMoCKb/PJdmdCeBbjZF1V7GnkgMsB6QgPyxvNveF8A11gzfvmzBa8mwFSAxsPkIgOc67PzolRBHMk7NQHGfDWay+22L0YGAFBhbHUuO+cyeiEY4tCsw8l5n1kgcjioJnAQdUyUNuHJr0DD6rWelpVB9Pg9/w6iBY+xfIA5XMfSh3P9qd/rAgDoHF/8QOG1ME2EeD83mNwftBj6FcSM+xZwbR3mRVPxRsIEoQArp5i8u1Z3QUbOPpf+c80AAK12YgT4cQ32Z+yWRE34mgAnesUR5WXA/6jzD8eydPiHJ8TnZKoTAGB6d3IS5GBUdxV5Y6o/+X0UbeRRnBtIvK3wRW2ljM9MJt6Jum+kS+nGV7JpqegOK5XYlXmyY5+V3pDy03ml68XCqdwhr7z0tcblu/mPeovrAmh+6XSjxNw3UR0G0gCW3UDmqU5M8s755JaXKZyYxltavpEqCoxqXA+GgYS+AjHOMVT33zAH8MrLFE5O4yws3jK/Wqbwbe52c4C0wgEq1rEwj/AbhtDhd756S8sUTuWwt94HCuXZP1An4D6IdoRZhD4By9MhIO+3po7LYu4yizOXA80RfhHxhsI8ajfhC1MNGzNXnke0H9gBbK1R8TPCaVQm5heaP+PTruUwcd0/p/fvzTYviWkxrnevhdcE4GH97Rrrr4S6c7+N9l6pd8+7Gv8AiH09KJR8pisAAAAASUVORK5CYII=';
+  static const String marker2x =
+      'iVBORw0KGgoAAAANSUhEUgAAADgAAAA/CAYAAAC1gwumAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8YQUAAAUJSURBVHgB3Zs9U+NGGMcf5CpDEdJQGwoYZmDwVVA6UKXAkJbmji/AkU9w8AkCPcxxDZSxoUl3TgfVOQMFQ4GdliakYCYNkOfvrDTy+ll7tbvi5PvNCMvSatk/z8u+aBmjHNjY2Kg+PT1VoihafHl5qfClCT7KWrHO2NhYhz8fnp+f/yiVSq16vd6kwIxRICCKxbzlY4P+F+TCA4uus+DG+fl5nQLgJZBFTXBj3vPpDrmLMtHhY5+9oMGW7ZAjTgJzFqbT4eP47OxsjxzILBCuyOI+Un9M5U2HrfljVmtGWQrXarUPLO4zvb44UObf3UYbsjxkZUG4JCeP3/ioUgFAIuJji635MLTssAIsrvwVrTYIK5cdKLDA4mKGijQKVJnyCxVXXExLiRTd1ZhkON5+peKLAxXVVpGSdFFlqh0aHSozMzP/3N7eXug3+lxUxV2bRhB21Tfsqq2ea3ohlVRGEslVewSura29o9GIOxH000pDQo9A7jwzjRKKiK4hiUEo55sfyZOTkxMaHx9PvvMgmY6OjsSy29vbtLq6mnx/fHykzc1N8oUtucXTrWOcJxYMYb35+fkecSAtQGdhYaHnO55FHb5wsnmbnOMHZ07MusvkydLSUt81U6NxbXJy0qqOrCAWMevBeaQuvKcA6BaJkRptEmKqIyvcG2BlIRFYJU9gjampKfHe8vKy1TWAOnQ3d2QdPyJ07BTAPQfFji4e55J7xoRwU6YMbRFWvygAwxqVvj8skYRyU9ZWjTh7VikA09PTA++nG21yz5hAFkTPUIHARfLElBH1MogtlBtmwVDdBVOOOMF4r4pJjbm+vu67trKyIpZtt9tWdTqwCAuWyRMpZk5PT/uuwTWljv/w8NCqTgcmvC0ouROGXLCgbhlkT70syqAsnkkjjYocmMi0bCghudLV1VXPZ4zUYFNZECLZeAuUGhHH3+Xl5bDHkzJSzIZwU8Tg0LXFQUiNiF0Tn7rrDSqrE8CCD4hBZ4HSiCSOv/hcanhMOvakOIRLm4Z/NuD1HCzYIkdsUv7FxYXxef1eDt1F14J/kSPSiERPFlJsme7d3d1Z/Q5bWNufEOhkQdNoQ2+0KQ7v7+/7LCYlJZ/ZBU98WxFeHZMDJteR3ExyU8la0rOecdiKsI7okkmlDGeyluSmkrVMScklmyLBQFs84f1EGZG6B6mzBpIYU2xKdTjGYRM/uqtq6q3tyC74SqgXMs1k2XB9ff3vEDOLIgD3bDQa3cBNhmos7oC+EVhgsmEhEcgm3fcdthUBtbmoGX9PXp/d3Nz8Ozs7+x2fVmmEYU/c49X03+PvPbMJZcUOjShoOy/Z76ev9QjEa2DsXqARRWp73xtedtUOu+oPfOo+CPw67HHmPNYvGjchcLfxRe0ULDyYEbG4N9K9aMBDP49CPKKNaKvx/qCHsfTNVvzMR5kKiBLntk8mpqgibcR1y5EFRRNpKw5YraqhIq4QQZx51pEDn9AW222VLvtF37ElP7y2NTGM5BnPnt6RD6NEGeF+sjU3N9dggegrX6UbYXFNPn5KD8GsnyUP2Jo7ypq5TLNcrdZTB3mSVwJSVtvy2ZDerYcCUavVdvkjyEYi/mP94mO1NMEEAl9rxqMSfUOdD0EFxjha84Cna7s2+7CzkItAYGtNZbWtPP6tB3i/PjORGhwMWus5UJ12k3IiNwum0QcHeVstzasIBGon8S4L+972fx5C8B/4bogVP/glegAAAABJRU5ErkJggg==';
+
+  static const String marker4x =
+      'iVBORw0KGgoAAAANSUhEUgAAAHAAAAB+CAYAAAD86pU7AAAACXBIWXMAACxLAAAsSwGlPZapAAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8YQUAAAn8SURBVHgB7Z0/cBNHFMafjxTJUERpIF0EBQ4MzDgN0EVAkzS26TI0QJdUIV1mKMBdUoFrCqBLFxs3pEJ0UOEMFDYFiI6hMgVFCiDvU3ad81l39/b+7L6T9zcjLInzSd5P39v33p32ZqijLC4uzr1//35uZmamz7ev+Knex48f58x/98wtzZa98fb4OeLtX/Ftfd++faOVlZV16iAz1AFYrB6LNeCB/5YfDvjWp90C1Yb3P4SgfHvIog5Z1C1SjloBWbT+hw8fLvKgDnhABxQACMrvYZXFZC1XRqQQVQIap11KkmQhlGgFwJnLxpkjUoIKAVk4uOwi3xaphdDYNOzMO3xb0iBkUAGNcNcUuk0EQqwRckiBCCJg14XLYoS8HMKRXgVEYmKEu0RTSIjQuo88weL9zBndH3z3NE0vc/hwHjly5J/nz58/Ig+07kDjutvTEi4dGHE2faZtN7bqwJTrvqa9R8+HG1txIOo5fvM3pnWuc8XMjb+00dlpXEDTQXlA/7W7Iv/TSkhNqEHQYI7i5TL+YGOMqEEamwPPnz9/0cx3X1IkD0wtPxw7duz1xsbG39QAjQgI8biHeYfvfkqRMj5Fy5BFHDUhYm0BU+JFHGhKxFoCRvHq0YSIlbNQ9DNNwhKpiclOh1SBSgKaUuEJdeDQT0fYYhG/qVJiOJcRqTovitccPVNi9MkRZwH5hf6kWOe1wbhnTI44JTHz8/PX+McPFGmL/uzsbG9zc/Mv6S+I58CYtPjDJakRCRj7m94ZmaSmtPktmgNxFJ2ieD6BYa5JNix1ILtv0SQuEc9IQmmpA1m8GxQJAo/9bRxbLdqmUECTdfYpEgqE0itFG+SG0Ji4qAFdmkN5CU2uA2PiooZekQsnOtC47yVFtJDrwk8mbW3cp4KbN2/SoUOHCre5cOECvXv3jlw4cOAA3bp1q3CbZ8+e0dWrV0kB1oXXs/+xK4SaM8oGpAAMcpl44OzZs+TK8ePHRdvs37+flPDzpIx0l4B8gBbfEOqTAiSDDE6fdj/Z+9y5c6LtTp06RUoYf/Uu++QuAWdmZtSET6mAcKmLU7CtdN8nTpwgLbA2F7PP7RAQDWtSlHlKB89FEOCyrSIHgjmj0TY7BMSXLEkJcBXmQCkuYdRFFNcPR9twMrOYfpwVcEBKcB00F1Fcw6ImAZkdJtsWUFv4dE1MpE7BNi7OBprmQaaXDqPbAmoKn1XDlsSFVUsOReXEjjCaFnBASqgasiSureomZWF0wd4ZC2jOhuqTEqpmfgiNRQPtmhilUSZg357BNhYQqyCRIurMOUUDXackkBb+vrCajQXkRum3pIQ6LgFF4lfp2FgwBx48eJC0YDUbC5haJC44dUNVXsIh7asWcfLkSdKCzVkS0yBVI2Adl1gmZZpNzGFNvLcGwTzYgwPViAfaGugm5jDXnqsH+glPhmq+49BUppcNo021w7CfumG4SbBeasIdbjUOrFJk55HOOJtsSGtqbmOxW4TQPinh8OHDpdu8efNmfCsjnY1KBh37lBzV1zQPYqXixCxXHBxplvjixQt69Kh83RyIZsOopK7E6RNPnz4t3Q7vU0s5wZnoFwn/o2IOlM5RGGjcyrDzlbSPiQ+FZL9ASzkB8+GkJhUCSucWDLIkhNp9SrPGly9fik+MQhhdW1uj0MB8n7CKOImJQiMJcxhgDDSAkGWuRekgEQX7tHMrti8TXVEm2lMRQqVhzooHJPMV9ilpy6X3lX6Nov0qaW73Gl1qqyrS8JkeaOl8JeHx48cTX6MILeWECgGlRx/SouG+68m8k8A+svuVoOUofXABXZrM2cGVhLsysvuQ7lNLWy24gC7lQxZJPVhGdh/pRKkMDWEUAo4oINJBQAGfJT13VWXSB0M6DyoIo6PgDpS0z8CkgZa21fLA705ym3QeDO1AXMQLrbQRBcLlFL+8Qa0TRvP2KRVQQTmxhTrwFQVC+scXdUnqJDJ54uO1pM4OKSBr9xYhNNgl1qRzSJFIdebBov12ZB4cwYFBLnzoEn6KBjNbx0mx7bM8pPsMedIvazdKcPVKCoBL6CkbzEkZahllDnP5UIRKZli7dYTQIA6U/tGSTLNKGC37HekBXhAwjK4n5ovzI/KM9I+WuMvlUBCQhl3NfVFUD9AuMQ+8utDl5F3JQLt0T4B0W5dywvchJp7/xutsWwEfkkeanP8sLvWgdFuXedB3OcECDvHTdmK8OlB6YpCLs1zmQakwrkfpfYIEBj/HAmJFPHNt9dZxKR9cwqI06chrn9V9D56PTmzZVQzTvdAhecAl1EiTCIskNLrWjNL34Hke3J7ytgXE9dLJAy4Zm+tgN5lZuuzT4isbTZJkZfu+vYOL3ZMHXGom1z6nZB50/VC4vAeP8+DQ3tmx2N3CwsKDPXip1E6BK2avrq6esY+TzH8uU0Q1rNHd9OPsAV1v2WjEHdN9uZN+boeAaM1wCI0u1Msw+8SuUyo4w7lDEZWwA5eyz+0SEFfQwkRJEVXgStiTrm6W5Gy8RBFV5GkyUUDTWhtSRAV57gNJwS9FFyqhSItcAU2z9C5FglLkPlB4Yi9npNdjXRgOjH1ZJCy8AOTGxsbW7OzsZ3x3QJEQ/M5ts8Ietej6gdwjfaJpOa69ALouLF7p8SnRdyN4Z5cp4hUe8zOS7UTX0OVQ+vro0aNv2YXfUcQHS2Wh0+J0Hfl4uKl9soeLynD6ehlCachvM007GFvX6crJgSBezbo9zIWPnc4QdLqOPOD5cBTnw+bh8fzl3r17zqe1OAsIWMRHXB/CvQOKNMHS2trab1SBSgKCzc3NIYv4Bd9VtYxtB1lm5/1KFaksIGAR77OIKDZjkV+Nuyzej1SDWgICFnElilgJiHeJalJbQGBEjHOinOW6zrM0IiAwc2IUsZylOnNelsYEBBAxlhiTGa/pkiQ/cZflJjWIcyEvAdf1YRHRdutTxHZYzrsW6RIadaAFxxHZiatY05n2eHKD3iaOLBQdVa+1f2oZduMVXJdey9rcvkDI5JYjCvRGQ+au1yEPmJB6e68cyTCuu9yW63a8Fnlk2t3oy3U7XpM8M60JDlb6MInKiDziXUDL/Pz8df5xjaYDFOZXKADBBARdd6M9AGsXHAjyHkgBHXXjMs6bNStdBUOFgKArbtTgujQqLjsAzNfavuG7mr9guoz3qEU8oMaBadiNl0y50ScFmFPc4TovK3m40EorrS7cilvX0opj4VZMKyzIspxlqHRgmlBuDFGUV0GlA9OEcKNphX3Ptd19Uo56B6ZpuxXXFdel6ZSAoK1yw2cDukk6J6ClyeIfJ9V2yXVpOisgqOvGNo+U+6LTAloqulFFK6wuUyEgkLpRWyusLmpaaXURtuLUtcLqMjUOTJMt/qfNdWmmUkCAkMo1HZZJ+dyIN5XLpfwLS6RBa1NRgoEAAAAASUVORK5CYII=';
 }
